LayoutTests/http/tests/iframe-monitor/workers/service-worker-not-count-twice-expected.txt

CONSOLE MESSAGE: setup() finished: true
CONSOLE MESSAGE: Service Worker registered with scope:
CONSOLE MESSAGE: Service Worker is ready and controlling the page.
CONSOLE MESSAGE: iframe is ready
CONSOLE MESSAGE: Get message from parent window: 5120
CONSOLE MESSAGE: Send fetch request to Service Worker
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/using-6k.html
Test ResourceMonitor doesn't count iframe's request twice. Iframe that using service worker can request regular resource fetch which might be through service worker. In this case we should't count those twice.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector('iframe[name=frame1]').srcdoc is ""
PASS document.querySelector('iframe[name=frame2]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame3]').srcdoc is ""
PASS successfullyParsed is true

TEST COMPLETE
